feat: Add prefinance projects tab to projects list (#2291)
