+++ web-components/src/components/cards/ProjectCard/ProjectCard.tsx
@@ -5,6 +5,7 @@ import { Buy1Event, Track } from 'web-marketplace/src/lib/tracker/types';
 
 import { ButtonType } from '../../../types/shared/buttonType';
 import { formatStandardInfo } from '../../../utils/format';
+import { cn } from '../../../utils/styles/cn';
 import OutlinedButton from '../../buttons/OutlinedButton';
 import BreadcrumbIcon from '../../icons/BreadcrumbIcon';
 import { PrefinanceIcon } from '../../icons/PrefinanceIcon';
