+++ web-components/src/components/header/components/HeaderMenuItem/HeaderMenuItem.styles.ts
@@ -4,6 +4,7 @@ export const useHeaderMenuHoverStyles = makeStyles()(theme => ({
   menuItem: {
     boxSizing: 'border-box',
     height: '100%',
+    fontWeight: 800,
     lineHeight: theme.spacing(6),
     paddingRight: theme.spacing(7.375),
     paddingLeft: theme.spacing(7.375),
