+++ web-components/src/theme/fonts.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css?family=Muli:800,900&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Muli:500,800,900&display=swap');
 @import url('https://fonts.googleapis.com/css?family=Lato:100,300,400,700,800&display=swap');
 
 @font-face {
