+++ web-marketplace/src/components/organisms/ProjectCardsSection/ProjectCardsSection.types.ts
@@ -1,4 +1,4 @@
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 
 export type ProjectCardOnButtonClickParams = {
   project: ProjectWithOrderData;
