+++ web-marketplace/src/components/organisms/ProjectTopSection/ProjectTopSection.types.ts
@@ -19,7 +19,7 @@ import {
   ProjectPageMetadataLD,
 } from 'lib/db/types/json-ld';
 
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 
 export type ProjectTopSectionProps = {
   offChainProject?: Maybe<ProjectFieldsFragment>;
