+++ web-marketplace/src/graphql/sanity/AllProjectsPage.graphql
@@ -3,5 +3,9 @@ query allProjectsPage {
     gettingStartedResourcesSection {
       ...gettingStartedResourcesSectionFields
     }
+    prefinanceProjects {
+      descriptionRaw
+      learnMore
+    }
   }
 }
