+++ web-marketplace/src/lib/normalizers/projects/normalizeProjectsWithMetadata.ts
@@ -15,8 +15,8 @@ import {
   ProjectPageMetadataLD,
 } from 'lib/db/types/json-ld';
 
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 import { getPriceToDisplay } from 'pages/Projects/hooks/useProjectsSellOrders.utils';
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
 import { getDisplayAccount } from 'components/templates/ProjectDetails/ProjectDetails.utils';
 
 interface NormalizeProjectsWithOrderDataParams {
