+++ web-marketplace/src/lib/tracker/useBuyModalOptionsTracker.ts
@@ -1,6 +1,6 @@
 import { Location } from 'react-router-dom';
 
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 
 import { BuyModalEvent, Track } from './types';
