+++ web-marketplace/src/pages/CreditClassDetails/CreditClassDetails.Projects.tsx
@@ -5,7 +5,7 @@ import cx from 'classnames';
 import { useWallet } from 'lib/wallet/wallet';
 
 import { BuySellOrderFlow } from 'features/marketplace/BuySellOrderFlow/BuySellOrderFlow';
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 import { ProjectCardsSection } from 'components/organisms/ProjectCardsSection/ProjectCardsSection';
 import { useProjectsWithOrders } from 'hooks/projects/useProjectsWithOrders';
