+++ web-marketplace/src/pages/Dashboard/MyProjects/hooks/useFetchProjectsWithOrders.tsx
@@ -9,7 +9,7 @@ import { normalizeProjectsWithOrderData } from 'lib/normalizers/projects/normali
 import { getSellOrdersExtendedQuery } from 'lib/queries/react-query/ecocredit/marketplace/getSellOrdersExtendedQuery/getSellOrdersExtendedQuery';
 import { useWallet } from 'lib/wallet/wallet';
 
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 import { useProjectsMetadata } from 'hooks/projects/useProjectsMetadata';
 
 interface Response {
