+++ web-marketplace/src/pages/Home/Home.FeaturedProjects.tsx
@@ -11,7 +11,7 @@ import {
 } from 'generated/sanity-graphql';
 
 import { BuySellOrderFlow } from 'features/marketplace/BuySellOrderFlow/BuySellOrderFlow';
-import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { ProjectWithOrderData } from 'pages/Projects/AllProjects/AllProjects.types';
 import { ProjectCardsSection } from 'components/organisms/ProjectCardsSection/ProjectCardsSection';
 
 import { useFeaturedProjects } from './hooks/useFeaturedProjects';
