+++ web-marketplace/src/pages/Projects/AllProjects/AllProjects.normalizers.ts
@@ -1,9 +1,12 @@
 import { AllCreditClassQuery } from 'generated/sanity-graphql';
 import { SKIPPED_CLASS_ID } from 'lib/env';
 
-import { CreditClassWithMedata } from './hooks/useFetchCreditClasses';
-import { UNREGISTERED_PATH, UNREGISTERED_PROJECTS } from './Projects.constants';
-import { CreditClassFilter, ProjectWithOrderData } from './Projects.types';
+import { CreditClassWithMedata } from '../hooks/useFetchCreditClasses';
+import {
+  UNREGISTERED_PATH,
+  UNREGISTERED_PROJECTS,
+} from './AllProjects.constants';
+import { CreditClassFilter, ProjectWithOrderData } from './AllProjects.types';
 
 type NormalizeCreditClassesFilterParams = {
   allProjects: ProjectWithOrderData[];
