+++ web-marketplace/src/pages/Projects/AllProjects/index.ts
@@ -1,0 +1,1 @@
+export { AllProjects as default } from './AllProjects';
