+++ web-marketplace/src/pages/Projects/PrefinanceProjects/index.ts
@@ -1,0 +1,1 @@
+export { PrefinanceProjects as default } from './PrefinanceProjects';
