+++ web-marketplace/src/pages/Projects/hooks/useOffChainProjects.tsx
@@ -5,7 +5,10 @@ import {
 } from '@apollo/client';
 import { useQuery } from '@tanstack/react-query';
 
-import { AllCreditClassQuery } from 'generated/sanity-graphql';
+import {
+  AllCreditClassQuery,
+  AllPrefinanceProjectQuery,
+} from 'generated/sanity-graphql';
 import { normalizeProjectWithMetadata } from 'lib/normalizers/projects/normalizeProjectsWithMetadata';
 import { getAllProjectsQuery } from 'lib/queries/react-query/registry-server/graphql/getAllProjectsQuery/getAllProjectsQuery';
 
@@ -14,11 +17,13 @@ import { findSanityCreditClass } from 'components/templates/ProjectDetails/Proje
 type Props = {
   sanityCreditClassesData?: AllCreditClassQuery;
   enabled?: boolean;
+  prefinanceProjectsData?: AllPrefinanceProjectQuery;
 };
 
 export const useFetchAllOffChainProjects = ({
   sanityCreditClassesData,
   enabled = true,
+  prefinanceProjectsData,
 }: Props) => {
   const graphqlClient =
     useApolloClient() as ApolloClient<NormalizedCacheObject>;
@@ -32,23 +37,31 @@ export const useFetchAllOffChainProjects = ({
   );
 
   const onlyOffChainProjectsWithData =
-    offChainProjects?.map(project => ({
-      offChain: true,
-      ...normalizeProjectWithMetadata({
-        offChainProject: project,
-        projectMetadata: project?.metadata,
-        projectPageMetadata: project?.metadata,
-        programAccount:
-          project?.creditClassByCreditClassId?.accountByRegistryId,
-        sanityClass: findSanityCreditClass({
-          sanityCreditClassData: sanityCreditClassesData,
-          creditClassIdOrUrl:
-            project?.creditClassByCreditClassId?.onChainId ??
-            project?.metadata?.['regen:creditClassId'] ??
-            '',
+    offChainProjects?.map(project => {
+      const prefinanceProject = prefinanceProjectsData?.allProject?.find(
+        sanityProject =>
+          sanityProject.projectId === project?.id ||
+          sanityProject.projectId === project?.slug,
+      );
+      return {
+        offChain: true,
+        ...normalizeProjectWithMetadata({
+          offChainProject: project,
+          projectMetadata: project?.metadata,
+          projectPageMetadata: project?.metadata,
+          programAccount:
+            project?.creditClassByCreditClassId?.accountByRegistryId,
+          sanityClass: findSanityCreditClass({
+            sanityCreditClassData: sanityCreditClassesData,
+            creditClassIdOrUrl:
+              project?.creditClassByCreditClassId?.onChainId ??
+              project?.metadata?.['regen:creditClassId'] ??
+              '',
+          }),
+          projectPrefinancing: prefinanceProject?.projectPrefinancing,
         }),
-      }),
-    })) ?? [];
+      };
+    }) ?? [];
 
   return {
     allOffChainProjects: onlyOffChainProjectsWithData,
