+++ web-marketplace/src/pages/Projects/hooks/useProjectsSellOrders.types.ts
@@ -1,10 +1,12 @@
-import { ProjectWithOrderData } from '../Projects.types';
+import { ProjectWithOrderData } from '../AllProjects/AllProjects.types';
 
 export interface ProjectsSellOrders {
   allProjects: ProjectWithOrderData[];
+  prefinanceProjects: ProjectWithOrderData[];
   projectsWithOrderData: ProjectWithOrderData[];
   projectsCount?: number;
   loading: boolean;
   hasCommunityProjects: boolean;
   haveOffChainProjects: boolean;
+  prefinanceProjectsCount?: number;
 }
